+++ src/main/java/org/jabref/gui/fieldeditors/JournalEditor.java
@@ -11,7 +11,7 @@ import org.jabref.logic.integrity.FieldCheckers;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.Field;
-import org.jabref.preferences.JabRefPreferences;
+import org.jabref.preferences.PreferencesService;
 
 import com.airhacks.afterburner.views.ViewLoader;
 
@@ -20,7 +20,11 @@ public class JournalEditor extends HBox implements FieldEditorFX {
     @FXML private JournalEditorViewModel viewModel;
     @FXML private EditorTextField textField;
 
-    public JournalEditor(Field field, JournalAbbreviationRepository journalAbbreviationRepository, JabRefPreferences preferences, SuggestionProvider<?> suggestionProvider, FieldCheckers fieldCheckers) {
+    public JournalEditor(Field field,
+                         JournalAbbreviationRepository journalAbbreviationRepository,
+                         PreferencesService preferences,
+                         SuggestionProvider<?> suggestionProvider,
+                         FieldCheckers fieldCheckers) {
         this.viewModel = new JournalEditorViewModel(field, suggestionProvider, journalAbbreviationRepository, fieldCheckers);
 
         ViewLoader.view(this)